!!! FEATURE: Add Spreads
ATTENTION: The changes the way props are returned, the props and spreads are now all returned as `attributes` of an
afx-node. The attributes can be of type `string`, `expression` and `spread` the first two contain an `identifier` with the
name of the prop. Spreads do not receive an identifier but only the expression as payload.